Implement geometry from config parameters
Implementation of using the configuration parameters for the
setup geometry. The target position is the now the center of the
laboratory reference frame. The CDS is then placed depending of
the relative position from the target position. The CDH is included
within the CDS therefore does not have independent positioning
anymore. Correct relative positioning of the internal layer of
CDS, before it was properly placed because the default position was
(0,0,0) and conflicts were not visible.