Add EventAction, RunAction and RunData classes
RunAction class create the run object that is use within the Geant4
state machine. It create a RunData object that holds the
ROOT objects: TFile, TTree and user classes. RunAction is
responsible for initializing the RunData with list of the sensitive
detectors, and for closing the run (the ROOT file mainly).
RunData class create and hold the output ROOT file defined in the
configuration, and the TTree of the simulated data. For each
sensitive detector, a branch of TClonesArray is create. FillEvent
is called by the EventAction after the end of each simulated event.
EventAction class create the connection between the Geant4 event
and the filling of the ROOT TTree.