Add default macro and config file
Add geant4 macro file, runTemplate.mac, for running the simulation
from the command line. It is just simple macro that can be used as
a base to include any modifitions. Please create your own macro
file, yet in case this is a start. Add also a configuration file
example testconfig.par that is the default config file that the
code require if no config file is provided in the command line.
Add dependency of the runTemplate.mac file to CMakeLists to
copy it into the build directory.