Add simplify geometry for ROOT
ROOT does not have a TwistedTubs equivalent. For the reconstruction
to have each cells of CDS is not necessary, yet the boundaries of
each layers is needed. Therefore, a simplified version of the
geometry is created when the conversion to ROOT geometry is
requested in the configuration file. Those extra volumes created
for this purpose are not existent when the simulation run without
the conversion to minimize the memory used.