Move .gitlab-ci.yml to diseable CI
Move .gitlab-ci.yml to .gitlab-ci.ymlNOYET to diseable the
Continous Integration provided by gitlab. The script need a
proper docker image that will be needed to be built and put
into the docker hub to be retrieved properly each time.
The different tests to find a proper docker image with ROOT
and Geant4 were unsuccessful. A dedicated docker will be built
later.