Add WASA geometry construction class
Add WASA geometry construction class in which the geometry is loaded from the ROOT file/gGeoManager. Use VGM for the translation from the ROOT geometry to Geant4 geometry. Additional detectors are created for the HypHI experiment purposes. Geometry is updated for the need of HypHI experiments. Field is handled. Sensitive detectors are set. Detector Cuts are also set.