Correct setting of magnetic field
Fix bug in the magnetic field class for setting the field. Fix the volume in which is the magnetic field. Right now there is no field map, just a constant field within the solenoid. Therefore the volume where there is a field is the INNER volume not the MFLD, which corresponds to the experimental area volume.