Add more options for driver in the magnetic field
Add more options for the setting of track driver in the magnetic field. Especially the difference from the defaults, is the use of  G4NystromRK4 stepper which is described as a faster stepper than G4ClassicalRK4 stepper for the same accuracy (but only usable for pure magnetic field, no electric field).