Fix initialization of PrimaryGeneratorAction for  Ions
Fix the initialization of the SimplePrimaryGeneratorAction 
for ions defined in the configuration file. If the config
file set the primary particle beam as a ion (Z...A...),
during the initialization the ion will be not found
because the Geant4 physicslists are not yet set for
ions. It leaded to segmentation fault. By defining a
generic particle like a geantino, it alllows a delayed
setting of the particle beam that is reset in
GeneratePrimaries when all physicslist defnitions are set.