Correct bug on the hit id of the central strawtube detector.
Correct bug that missed to label the id of the wire in each layer
of the central strawtube detector. The other detectors are fine.
The problem occurred because the geometry of the CDC is as such,
layer (MD..), wire (ME..), and inner wire (MG..). The inner wires
are not replicates and has the id 1, while the id of wire is on ME.
The wire (ME..) can not be accessed as the current physical volume,
but as the mother physical volume, which is normally the case with
other detectors. By testing the depth of the touchable history,
it is possible to distinguish between the CDC and the other
detectors. As such, if it the CDC i.e a wire id, a mother physical
volume information is used, else it current volume information
that is accessed.