Add condition for Fiber, Si and virtual trackers activation
Add condition for the fiber, Si, and/or virtual tracker activation.
Via the config file, each section can be set for the simulation.
Instead of commenting the existing code, conditioning its use allow
more flexibility. Some of the constant of the fiber definition are
set as const to make sure that they are not modified.