Add explicit define variable for G4VIS_USE and G4UI_USE
Add explicit #define for G4VIS_USE and G4UI_USE. It seems that the
project configuration in cmake does not propagate those variables
anymore. Without those variable, the GUI interface of geant4 does
not start. For quick fix those variable are explicitly defined at
the start of the source file, and the GUI works again.