Fix the PSB and PSF segmentation in geometry file of WASA
Fix the PSB and PSF segmentation into PSBE and PSFE in the geometry
file of the WASA collaboration. The member function Divide were
used to produce the segmentation of PSB and PSF and the placement
of PSBE and PSFE. This function did produce correct placement inside
of ROOT, but once loaded in the geant4 framework, the placement were
not correct which leaded to wrong geometry for PSB and PSF, and
the resulting simulations were incorrect for PSB and PSF. Some hits
were not produced, and the global hit position were not related with
the correct segment, and so wrong hit measurement per segment.
The problem is fixed by passing the volume segment of PSFE and PSBE
manually. All analysis before this correction involving PSBE or PSFE
can not trusted.