Add possibility for elliptical beam spot in primary generator
Add the possibility to generate events from a elliptical beam spot
in the primary generators. Up to now, the beam spot was Gaussian
distributed in the radial direction, generating a 2D-Gaussian with
same variance in X and Y direction. Now the case of different
variance in X and Y is possible. The option needed in the
configuration file is Beam_SpotSizeSigmaX and Beam_SpotSizeSigmaY.
They need to be set both.