Import magnetic field of WASA solenoid from the WASA software
Import the MField class from MField.cc / MField.hh of WASA software.
The MField comes from wasa/WasaParameter. The MField class was
adapted to be used outside of WASA software. The file of default
magnetic field map is also included. The field map is represented
as a function of the radial and longitudinal (R, Z) positions for
radial and longitudinal magnetic field (in kGauss).