Add Geant4 magnetic field class to include the WASA field map
Add new G4 class for the magnetic field of the WASA field map.
Modify the WasaDetectorConstruction to allow selecting between
the simple uniform field using "Field_CDS_Bz" config option
and the field map via the "Field_WASAMap" config option.
The maximum of the field can be set via "Field_WASAMapMaxField"
config option. Additionally, the origin position of the field
needs to be given and set.