Add option to minimize the physical volume visualization in the GUI
Add the config option "Gui_MiniVis" in order to minimize the physical
volumes in the geometry for the graphical interface. When all the
fiber channels, all the silicon strips, all the strawtubes are visible
it can be difficult to navigate around the 3D model in GUI. With this
option the mother volumes are made visible while hiding the underlying
daughter volumes: individual fibers, strips, strawtubes are no
visible. The simulation behavior is still perfectly normal, the
sensitive volumes are still sensitive. It is just about the geometry
representation in the graphical interface.