Add the CsI ball into the ROOT geometry file
Add into the ROOT geometry file the CsI ball in for the EM
calorimeter. First the EM can be now be selected between the central,
front and back-end. Second, the sides of the crystals are not
perfectly co planar, which lead to problem in geant4. cf:
https://geant4-forum.web.cern.ch/t/g4trap-coplanarity-problem-with-geant4-version-10-4/1831
There is a function to correct the geometry of the crystals.
Included the TrapParamsAdjustment function for this purpose.