Change the WASA standard geometry for more flexibility
Change the WASA standard geometry to be more flexible. The World
volume "WASA" is made bigger in case that the MFLD volume is moved.
The INNER volume in which the solenoid sub volume was defined is
changed in order to represent only the inner detectors (MDC + PSB).
The solenoid volume is not at the same level as the INNER volume,
so that they can be moved and adjusted independently. Finally the
volumes are made of air instead of vacuum