Modify the detector constructionto adjust to the WASA-FRS experiment
Modify the detector construction in order to adjust to the experiment.
The target material was diamond so it is now possible to choose it.
The mini fiber detectors are defined as to separated volumes so that
they can be placed independently as in the experiment. The FMF2 and
EndCap detectors are now optional.