Add new detector construction based only on the ROOT geometry
Add a new detector construction class to create a geant4 geometry
based only from the ROOT geometry. The normal WASA geometry setup,
is based on the WASA ROOT file and the geant4 detector definition.
Once the geometry is defined and set from the experimental setup,
the geometry must not be modified anymore, therefore with this new
detector construction the geometry is loaded without modification.
The geometry controller is naming the geometry "WasaFullRoot", used
to select this construction.