Add ParsePreviousParams function in the config class
Add the member function ParsePreviousParams in the config class.
It is used to parse the data structure saved in the ROOT geometry
file. The parameters are then reloaded into the configuration
data structure in order to make sure that the used parameters are
re-used correctly and to keep everything consistent.