Fix the rotation of MDC & PSCE for the calibration input parameters
Fix the rotation of tMDC and PSCE volume from the values obtained
from the calibration of the experimental geometry. The
G4RotationMatrix must exist during all the run of simulation since
we are passing a pointer to the SetRotation member function of
G4Volume. If the RotationMatrix object is create locally in the
Construct() function or in local scope, the object will be destroyed
when it goes out-of-scope, and won't exist when needed in the
construction or use of the geometry.