Modify the import of the spatial calibration for the Experiment simu
Modify the import of the spatial calibration parameters of the fibers,
MDC and PSB that was located in the file ExpParams/WASAFRS_Calib.dat.
The parameters are now in the same repository of the other calibration
parameters, in such way that everything is located into the same place
for the simulations and the analysis. In ExpParams/WASAFRS_Calib.dat
and ExpParams/WASAFRS_Experiment.dat, the calibration parameters are
first commented. They will be removed in later commits. The
modification of SolConfig and WasaDetectorConstruction are for reading
the new files in which are those parameters.